Alter proceedings_before_the_courts to use application proceeding types
The application proceeding type is sent as an id to the proceedings_before_the_court.
This is all because the chances of success model is now under the application
proceeding type.

The legal aid application is set as an instance in the controller because of the ProviderBaseForm requiring it for authorisation checks.

Update tests

Alter the form to not require validation when save as draft button is selected similar to other pages on the app.

Update the locales