Update chances of success factory
Update the chances of success factory to have a trait for application proceeding types instead
of a default application proceeding type. This is because in some tests we create application proceeding types
seperately and pass in the chances of success. This causes a factorybot stack level too deep because the
application proceeding type creates a chances of success which creates an application proceeding type by default.

This is a workaround the stack level too depp bug.