Refactor the ccms base submission service
DRY out the ccms base submission service by removing some methods.

Remove the sentry calls because Sidekiq will need the exceptions to be raised in order to message Sentry and
to retry the worker.

Remove the submission.fail! because we are no longer failing a submission as it is being retried until exhausted by Sidekiq.
At that point, sidekiq will consider it a fail and the worker will call submission.fail!. No need for the inbetween submitters to handle this.

Add a new ccms error that can handle storing the response details when an exception is raised alongside its response as a variable.
Include these exceptions as ones to be caught and re-raised to sidekiq in the worker.