Update reports tests to include ccms submission information
The reports services require the ccms case reference and if it does not exist then it creates it.

Add ccms_submission with and without case reference tests to the existing reports tests.

Allow for RSpec::Mocks.configuration.allow_message_expectations_on_nil to be true because
the expect(nil).to receive(:process!) was trying to create a ccms_submission which was failing.